refactor(core): merge BaseApi into ApiClient
BaseApi existed only to be extended — it kept every verb protected, so the two
classes that wanted a client had to subclass it just to expose them. That split
is also what let ApiClient ship overrides whose signatures did not match the
parent, which PHP rejects at class-load time and php -l cannot catch.

One class now. The constructor takes either a connection id or a ready
AuthStrategyInterface, so ConnectionTestApi passes its handler exactly as before
and needed only its extends clause changed. BaseApi::request() became a private
dispatch(); send() is the entry point and request() stays as an alias.

Verified by loading both classes and exercising both construction paths,
ConnectionTestApi's success/rejection/empty-endpoint result shapes, and the
cross-app connection guard. Not verified against a live connection: the real
OAuth1/OAuth2/Basic/Bearer strategies have only gone through stubs.
