+++ backend/Authorization/Exception/AuthorizationException.php
@@ -11,7 +11,7 @@ use Exception;
 /**
  * Thrown by auth strategies when credentials cannot be produced (missing
  * fields, decrypt failure, invalid handler config). Must never escape the
- * Flow engine: BaseApi::request() converts it to ApiResponse::failure(),
+ * Flow engine: ApiClient::send() converts it to ApiResponse::failure(),
  * and CredentialInjector::inject() catches all Throwables.
  */
 class AuthorizationException extends Exception
