+++ backend/Core/Http/ApiClient.php
@@ -7,7 +7,11 @@ if (!defined('ABSPATH')) {
 }
 
 use BitApps\Integrations\Authorization\AuthorizationFactory;
+use BitApps\Integrations\Authorization\Contract\AuthStrategyInterface;
+use BitApps\Integrations\Authorization\Exception\AuthorizationException;
+use BitApps\Integrations\Authorization\RequestContext;
 use BitApps\Integrations\Core\Database\ConnectionModel;
+use BitApps\Integrations\Core\Util\HttpHelper;
 use Throwable;
 
 /**
@@ -24,19 +28,40 @@ use Throwable;
  * The auth type and app slug come from the connection row, so nothing but the id is
  * needed. Credentials resolve on the first request, not on construction: building a
  * client stays free of side effects (an OAuth2 getter can trigger a token refresh).
+ * A ready AuthStrategyInterface can be passed instead of an id, for callers that
+ * already hold one.
  *
  * setBody() sets the payload for the NEXT request only — it is cleared once sent, so a
  * client reused across calls never leaks one request's body into the next.
  *
- * Every call returns an ApiResponse (ok/status/body/error); what counts as a failure is
- * the caller's decision, since providers disagree about how they report one.
+ * Contract notes:
+ * - credential() is fetched per call and never memoized: strategies may compute
+ *   per-request values (KirimEmail HMAC + Timestamp).
+ * - The HTTP status is captured immediately after the request, before any nested call
+ *   can clobber the HttpHelper::$responseCode static.
+ * - Multipart/CURLFile bodies are untested on this path.
  */
-class ApiClient extends BaseApi
+class ApiClient
 {
+    /**
+     * @var null|AuthStrategyInterface
+     */
+    protected $auth;
+
+    /**
+     * @var array<string, string>
+     */
+    protected $defaultHeaders = [];
+
     /**
      * @var int
      */
     private $connectionId = 0;
+
+    /**
+     * @var null|string when set, the connection must belong to this app
+     */
+    private $appSlug;
 
     /**
      * @var mixed payload for the next request
@@ -54,24 +79,44 @@ class ApiClient extends BaseApi
     private $baseUrlOverride;
 
     /**
+     * @var null|string resolved endpoint base, cached so an OAuth2 getter runs once
+     */
+    private $resolvedBase;
+
+    /**
+     * @var bool
+     */
+    private $baseResolved = false;
+
+    /**
      * @var null|ApiResponse the most recent response, for callers that inspect it
      */
     private $response;
-
-    /**
-     * @var null|string when set, the connection must belong to this app
-     */
-    private $appSlug;
 
     /**
      * @var null|string why the client could not be prepared
      */
     private $setupError;
 
-    public function __construct($connectionId = 0)
+    /**
+     * @var null|AuthorizationException set by the last request that failed to build a credential
+     */
+    private $lastAuthException;
+
+    /**
+     * @param AuthStrategyInterface|int $connection a connection id, or a ready handler
+     */
+    public function __construct($connection = 0)
     {
-        if (!empty($connectionId)) {
-            $this->setConnectionId($connectionId);
+        if ($connection instanceof AuthStrategyInterface) {
+            $this->auth = $connection;
+            $this->authResolved = true;
+
+            return;
+        }
+
+        if (!empty($connection)) {
+            $this->setConnectionId($connection);
         }
     }
 
@@ -83,6 +128,8 @@ class ApiClient extends BaseApi
         $this->setupError = null;
         // The old base belonged to the old connection's tenant.
         $this->baseUrlOverride = null;
+        $this->resolvedBase = null;
+        $this->baseResolved = false;
 
         return $this;
     }
@@ -101,8 +148,8 @@ class ApiClient extends BaseApi
     public function setAppSlug(?string $appSlug): self
     {
         $this->appSlug = $appSlug;
-        $this->authResolved = false;
         $this->auth = null;
+        $this->authResolved = false;
 
         return $this;
     }
@@ -118,9 +165,13 @@ class ApiClient extends BaseApi
             return $this->baseUrlOverride;
         }
 
-        $auth = $this->auth();
+        if (!$this->baseResolved) {
+            $this->baseResolved = true;
+            $auth = $this->auth();
+            $this->resolvedBase = $auth === null ? '' : rtrim((string) $auth->getEndpointBase(), '/');
+        }
 
-        return $auth === null ? '' : rtrim((string) $auth->getEndpointBase(), '/');
+        return (string) $this->resolvedBase;
     }
 
     public function setBaseURL(?string $baseUrl): self
@@ -207,8 +258,6 @@ class ApiClient extends BaseApi
      */
     public function send(string $method, string $path = '', $payload = null, array $headers = []): ApiResponse
     {
-        // setBody() is for one request; taking it here stops a reused client from
-        // sending a previous call's payload.
         if ($payload === null) {
             $payload = $this->body;
         }
@@ -219,14 +268,20 @@ class ApiClient extends BaseApi
             return $this->response = ApiResponse::fail(0, $this->setupError ?: __('Connection is not configured', 'bit-integrations'));
         }
 
-        return $this->response = parent::request($method, $path, $payload, $headers);
+        return $this->response = $this->dispatch($method, $path, $payload, $headers);
     }
 
     /**
-     * The most recent response. Platforms report failures differently — some inside a
-     * 2xx body — so the caller decides what counts as an error rather than the client
-     * guessing on its behalf.
+     * Alias of send(), kept for callers that read as "make this request".
+     *
+     * @param null|array|object|string $payload
      */
+    public function request(string $method, string $path = '', $payload = null, array $headers = []): ApiResponse
+    {
+        return $this->send($method, $path, $payload, $headers);
+    }
+
+    /** The most recent response, for callers that judge failure from the body. */
     public function getResponse(): ?ApiResponse
     {
         return $this->response;
@@ -253,17 +308,105 @@ class ApiClient extends BaseApi
         return $this->setupError;
     }
 
-    protected function baseUrl(): ?string
+    /**
+     * The AuthorizationException from the most recent request, when the strategy could
+     * not produce a credential and no request was sent. Distinguishing that from a real
+     * transport/HTTP failure by inspecting the response body alone is not reliable, so
+     * the signal is explicit.
+     */
+    protected function lastAuthException(): ?AuthorizationException
     {
+        return $this->lastAuthException;
+    }
+
+    /**
+     * @param mixed $raw decoded JSON (object/array/scalar) or raw body string
+     */
+    protected function normalize($raw, int $status): ApiResponse
+    {
+        if ($status >= 200 && $status < 300) {
+            return ApiResponse::ok($status, $raw);
+        }
+
+        return ApiResponse::fail($status, null, $raw);
+    }
+
+    protected function resolveUrl(string $path): string
+    {
+        if ($path === '' || preg_match('#^https?://#i', $path)) {
+            return $path;
+        }
+
         $base = $this->getBaseURL();
 
-        return $base === '' ? null : $base;
+        if ($base === '') {
+            return $path;
+        }
+
+        return $base . '/' . ltrim($path, '/');
+    }
+
+    /**
+     * @param null|array|object|string $payload
+     */
+    private function dispatch(string $method, string $path, $payload, array $headers): ApiResponse
+    {
+        $this->lastAuthException = null;
+
+        $method = strtoupper(trim($method));
+        $method = $method === '' ? 'GET' : $method;
+        $url = $this->resolveUrl($path);
+        $headers = array_merge($this->defaultHeaders, $headers);
+
+        // The URL and method are resolved BEFORE the credential is built: an OAuth1
+        // signature is computed over them, so a credential produced first would be
+        // signing a request that does not exist yet. Only array payloads are passed —
+        // those are the ones sent form-encoded, and an OAuth1 signature covers form body
+        // params but never a JSON or multipart body.
+        try {
+            $credential = $this->auth->credential(
+                new RequestContext($method, $url, \is_array($payload) ? $payload : [])
+            );
+        } catch (AuthorizationException $e) {
+            $this->lastAuthException = $e;
+
+            return ApiResponse::fail(0, $e->getMessage());
+        }
+
+        if ($credential->isQuery()) {
+            if ($method === 'GET') {
+                // HttpHelper::get emits $payload as the query string; merge auth
+                // params into it so one deduplicated query is sent. Caller keys
+                // win on collision (matches the legacy authorize() behavior).
+                $payload = array_merge($credential->data(), \is_array($payload) ? $payload : []);
+            } else {
+                $query = http_build_query($credential->data());
+                $separator = strpos($url, '?') !== false ? '&' : '?';
+                $url .= $separator . $query;
+            }
+        } else {
+            $headers = array_merge($headers, $credential->data());
+        }
+
+        $raw = HttpHelper::request($url, $method, $payload, $headers, $this->auth->requestOptions());
+
+        if (is_wp_error($raw)) {
+            return ApiResponse::fail(0, $raw->get_error_message(), $raw);
+        }
+
+        // Capture immediately: the static is rewritten by the next request anywhere
+        // in the process (nested calls, related actions).
+        $status = isset(HttpHelper::$responseCode) ? (int) HttpHelper::$responseCode : 0;
+
+        return $this->normalize($raw, $status);
     }
 
     /**
      * app_slug is stored as the integration's display name ("Zoho CRM") while callers
      * pass a bare slug ("zohocrm"), so both sides reduce to alphanumerics before
      * comparing. No slug set means the caller opted out of the check.
+     *
+     * @param mixed $storedSlug
      */
     private function belongsToApp($storedSlug): bool
     {
@@ -282,7 +425,7 @@ class ApiClient extends BaseApi
      * The connection row carries both the auth type and the app slug, so the handler is
      * built from the id alone. Resolved once, on first use.
      */
-    private function auth()
+    private function auth(): ?AuthStrategyInterface
     {
         if ($this->authResolved) {
             return $this->auth;
